cleaned up the checkpoints branch and modified paths
