+++ dvc-get-started/code-checkpoints/signal-file/src/train.py
@@ -4,11 +4,25 @@ import yaml
 import os
 
 import models
-from dvc.api import make_checkpoint
 
 MODEL_FILE = "models/model.h5"
 
-class DVCLiveCallback(tf.keras.callbacks.Callback):
+class DVCCheckpointsCallback(tf.keras.callbacks.Callback):
+
+    def __init__(self, frequency = 1):
+        self.frequency = frequency
+
+    def dvc_signal(self):
+        "Generates a DVC signal file and blocks until it's deleted"
+        dvc_root = os.getenv("DVC_ROOT") # Root dir of dvc project.
+        if dvc_root: # Skip if not running via dvc.
+            signal_file = os.path.join(dvc_root, ".dvc", "tmp",
+                "DVC_CHECKPOINT")
+            with open(signal_file, "w") as f: # Write empty file.
+                f.write("")
+            while os.path.exists(signal_file): # Wait until dvc deletes file.
+                # Wait 10 milliseconds
+                time.sleep(0.01)
 
     def on_train_begin(self, logs=None):
         pass
@@ -20,10 +34,8 @@ class DVCLiveCallback(tf.keras.callbacks.Callback):
         pass
 
     def on_epoch_end(self, epoch, logs=None):
-        logs = logs or {}
-        for metric, value in logs.items():
-            dvclive.log(metric, value)
-        dvclive.next_step()
+        if (epoch % self.frequency) == 0:
+            dvc_signal(self)
 
     def on_test_begin(self, logs=None):
         pass
@@ -74,7 +86,7 @@ def history_to_csv(history):
 
 def main():
     params = load_params()
-    if params["continue"] and os.path.exists(MODEL_FILE):
+    if params["resume"] and os.path.exists(MODEL_FILE):
         m = tf.keras.models.load_model(MODEL_FILE)
     else:
         m = models.get_model()
@@ -99,15 +111,13 @@ def main():
     print(f"y_train: {y_train.shape}")
     print(f"y_valid: {y_valid.shape}")
 
-    dvclive.init("training_metrics")
-
-    m.fit(x_train,
+    history = m.fit(x_train,
                     y_train,
                     batch_size = params["batch_size"],
                     epochs = params["epochs"],
                     verbose=1,
                     validation_data = (x_valid, y_valid),
-                    callbacks=[DVCLiveCallback()])
+                    callbacks=[DVCCheckpointsCallback(frequency=1)])
 
     with open("logs.csv", "w") as f:
         f.write(history_to_csv(history))
