--- dvc-get-started/code-cp-2-basic/logs.csv
@@ -1,11 +1,0 @@
-epoch, loss, categorical_accuracy, precision, recall, val_loss, val_categorical_accuracy, val_precision, val_recall
-1, 0.6911119818687439, 0.8150208592414856, 0.941104531288147, 0.6571458578109741, 0.3279571235179901, 0.9090833067893982, 0.9434317946434021, 0.8755833506584167
-2, 0.3182624876499176, 0.909541666507721, 0.9370669722557068, 0.8819166421890259, 0.2747580111026764, 0.921999990940094, 0.9450414776802063, 0.9013333320617676
-3, 0.2791164517402649, 0.9209166765213013, 0.9413698315620422, 0.9008125066757202, 0.25979068875312805, 0.9242500066757202, 0.9425605535507202, 0.9079999923706055
-4, 0.2577449679374695, 0.9270208477973938, 0.9441564083099365, 0.9105208516120911, 0.2423638254404068, 0.9295833110809326, 0.9467206001281738, 0.9165833592414856
-5, 0.2418537735939026, 0.9310833215713501, 0.9466056227684021, 0.9170833230018616, 0.22990162670612335, 0.9360833168029785, 0.9497595429420471, 0.921583354473114
-6, 0.22818522155284882, 0.9338750243186951, 0.9487674832344055, 0.9213125109672546, 0.22259455919265747, 0.9380000233650208, 0.951818585395813, 0.9268333315849304
-7, 0.21706406772136688, 0.937166690826416, 0.9508551359176636, 0.9254791736602783, 0.21552026271820068, 0.9392499923706055, 0.9522507786750793, 0.9290000200271606
-8, 0.2071533054113388, 0.9400208592414856, 0.95302814245224, 0.9290833473205566, 0.2088119089603424, 0.9409999847412109, 0.9535479545593262, 0.9305833578109741
-9, 0.19879712164402008, 0.942187488079071, 0.9547946453094482, 0.9324166774749756, 0.20639696717262268, 0.9424999952316284, 0.9553373456001282, 0.9322500228881836
-10, 0.19090032577514648, 0.945229172706604, 0.9566746950149536, 0.9347708225250244, 0.1979590505361557, 0.9443333148956299, 0.9546072483062744, 0.9358333349227905
