--- dvc-get-started/code-cp-2-basic/metrics.json
@@ -1,1 +1,0 @@
-{"loss": 0.19828975200653076, "categorical_accuracy": 0.9426000118255615, "precision": 0.9547034502029419, "recall": 0.9337000250816345}
