--- dvc-get-started/code-cp-2-basic/src/__init__.py
@@ -1,1 +1,0 @@
-
