--- dvc-get-started/code-cp-3-signal-file/src/__init__.py
@@ -1,1 +1,0 @@
-
