--- dvc-get-started/code-cp-3-signal-file/src/evaluate.py
@@ -1,27 +1,0 @@
-import tensorflow as tf
-import numpy as np
-import json
-import yaml
-
-def load_npz_data(filename):
-    npzfile = np.load(filename)
-    return (npzfile['images'], npzfile['labels'])
-
-def load_params():
-    return yaml.safe_load(open("params.yaml"))
-
-def main():
-    params = load_params()
-
-    test_img, test_labels = load_npz_data("data/preprocessed/mnist-test.npz")
-    model = tf.keras.models.load_model("models/model.h5")
-
-    metrics_dict = model.evaluate(test_img, test_labels, batch_size=params["train"]["batch_size"], return_dict=True)
-    metrics_file = "metrics.json"
-
-    with open(metrics_file, "w") as f:
-        f.write(json.dumps(metrics_dict))
-
-
-if __name__ == "__main__":
-    main()
